TreeArrowRenderer contains a static function to render the TreeStructur with its arrows
TreeNodeComponent interface for TreeNodes that get implemented in the TypeCheckerNodeComponent
and later in the BigStepNodeComponent.

TreeNodeLayout class handling the indentated Layout of a tree.

TypeCheckerComponent, TypeCheckerNodeComponent implementation of the TypeChecker gui

Main TestDialog. Just for testing purposes

ExpressionRenderer renamed to PrettyString renderer and is no able to do propper rendering.

CompoundExpression is no able to do propper rendering.

EnvironmentRenderer is no able to do propper rendering

git-svn-id: https://vcs.informatik.uni-siegen.de/svn/tpml/trunk@282 69bfd797-8304-0410-886d-9c04947c740f